Rfcs post and get

+++ src/main/java/dsd/api/cdmsa/repository/OrgRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dsd.api.cdmsa.model.Organization;
 
-public interface OrgRepository extends JpaRepository<Organization, Integer>{
+public interface OrgRepository extends JpaRepository<Organization, Long>{
 
     boolean existsByName (String email);
 }
